requirements.txt
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
python-dotenv==0.19.2
pymysql==1.0.2
appwrite==1.0.0
